// src/app/api/models/evento.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { FechasEventos } from '../models/fechas-eventos';
import { Inscripcion } from '../models/inscripcion';
import { Prioridad } from '../models/prioridad';
export interface Evento {
  asistentesRegistrados?: number;
  capacidadMaxima?: number;
  descripcion?: string | null;
  fechaHora?: string | null;
  fechas?: FechasEventos;
  id?: number;
  inscripciones?: Array<Inscripcion> | null;
  nombre?: string | null;
  prioridad: Prioridad;
  ubicacion?: string | null;
  usuarioCreadorId?: number;
  usuarioCreadorNombre?: string | null;
}